refactor(ButtonContainer): 컨테이너로 다시 바꾸고 ProgressContainer를 밖으로 꺼냄
