+++ src/features/apply-form/ui/index.ts
@@ -1,3 +1,3 @@
+export * from './containers';
 export * from './form';
-export * from './section';
 export * from './steps';
