--- src/features/apply-form/ui/section/ProgressSection.tsx
@@ -1,15 +1,0 @@
-import { Progress } from '@/shared';
-
-type Props = {
-  step: number;
-  progressValue: number;
-};
-
-export const ProgressSection = ({ step, progressValue }: Props) => {
-  return (
-    <div className='mb-3 flex w-2/5 items-center gap-3'>
-      <Progress value={progressValue} className='flex-1' />
-      <p className='text-sm font-semibold'>{step} / 2 단계</p>
-    </div>
-  );
-};
