--- src/features/apply-form/ui/section/index.ts
@@ -1,1 +1,0 @@
-export * from './ProgressSection';
