+++ src/pages/items/ItemsPage.tsx
@@ -1,7 +1,7 @@
 import { ItemsTableSection } from '@/features';
 import { BackButton, NAVIGATE_BUTTONS_DATA } from '@/shared';
 import {
-  ButtonSection,
+  ButtonContainer,
   ContentsContainer,
   DescriptionSection,
   PageWrapper,
@@ -16,9 +16,9 @@ export const ItemsPage = () => {
         </DescriptionSection>
         <ItemsTableSection />
       </ContentsContainer>
-      <ButtonSection>
+      <ButtonContainer>
         <BackButton />
-      </ButtonSection>
+      </ButtonContainer>
     </PageWrapper>
   );
 };
