+++ src/widgets/containers/ButtonContainer.tsx
@@ -1,0 +1,11 @@
+type Props = {
+  children: React.ReactNode;
+};
+
+export const ButtonContainer = ({ children }: Props) => {
+  return (
+    <div className='w-layout flex items-center justify-center gap-3 pb-6'>
+      {children}
+    </div>
+  );
+};
