+++ src/widgets/sections/index.ts
@@ -1,2 +1,1 @@
-export * from './ButtonSection';
 export * from './DescriptionSection';
